data/de-w0-starter/pyproject.toml
[project]
name = "de-w0-starter"
version = "0.1.0"
description = "Week 0 starter for DE Projects"
requires-python = ">=3.10"
dependencies = [
  "pandas",
  "python-dotenv"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "mypy>=1.17.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
]
